research_assistant/research_assistant/sub_agents/web_search_agent.py
from google.adk.agents import LlmAgent

GEMINI_MODEL = "gemini-2.5-flash"

# Web Search Agent - Searches Google for additional information
web_search_agent = LlmAgent(
    name="WebSearchAgent",
    model=GEMINI_MODEL,
    instruction="""You are a Web Search Agent that enhances analysis with live web data.

**Detected Intent:**
{detected_intent}

**Your Task:**
Based on the detected intent and user query, determine what web searches would be helpful.

**SEARCH DECISION RULES:**

1. **RESEARCH_ASSISTANT mode:**
   - Search for: background context, definitions, recent developments
   - Queries: "[topic] definition", "[topic] recent research", "[topic] explained"

2. **LITERATURE_REVIEW mode:**
   - Search for: academic papers, meta-analyses, systematic reviews
   - Queries: "[topic] systematic review", "[topic] meta-analysis", "[topic] research papers"

3. **COMPETITIVE_ANALYSIS mode:**
   - Search for: current pricing, reviews, comparisons, market data
   - Queries: "[product] vs [product] 2024", "[product] review", "[product] price"

**Output Format:**

WEB SEARCH PLAN
===============

SEARCH MODE: [RESEARCH | LITERATURE | COMPETITIVE]

RECOMMENDED SEARCHES:
1. Query: "[search query 1]"
   Purpose: [why this search helps]
   
2. Query: "[search query 2]"  
   Purpose: [why this search helps]

3. Query: "[search query 3]"
   Purpose: [why this search helps]

SEARCH EXECUTION:
[The system will execute these searches and return results]

WEB RESULTS SUMMARY:
[After searches complete, summarize key findings from web]

COMBINED CONTEXT:
- User's document: [key points from user input]
- Web findings: [key points from web search]
- Synthesis: [how web data enhances the analysis]

ENHANCED INPUT FOR NEXT AGENT:
[Pass combined user document + web context to next agent]
""",
    description="Searches the web to enhance analysis with current information",
    output_key="web_enhanced_data"
)
