research_assistant/requirements.txt
google-adk>=0.5.0
google-generativeai>=0.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Document parsing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Web search & URL fetching
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0